@@ -1,32 +1,28 @@
 {
-  "name": "portfolio",
+  "name": "matt-gainer-development",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "bootstrap": "^5.1.0",
-    "emailjs": "^3.6.0",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "emailjs": "^4.0.2",
     "emailjs-com": "^3.2.0",
-    "jquery": "^3.6.3",
-    "link": "^0.1.5",
     "particles-bg": "^2.5.5",
-    "react": "^16.0.2",
-    "react-bootstrap": "^2.0.0-beta.6",
-    "react-dom": "^16.0.2",
-    "react-external-link": "^1.2.2",
-    "react-ga": "^3.3.1",
-   
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "4.0.3",
-    "react-zmage": "^0.8.5-beta.37",
-    "sass": "^1.38.1",
-    "web-vitals": "^1.1.2"
+    "primereact": "^9.5.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.13.0",
+    "react-scripts": "5.0.1",
+    "sass": "^1.63.4",
+    "web-vitals": "^2.1.4"
+  },
+  "devDependencies": {
+     "prettier": "^3.0.0-alpha.6"
   },
   "scripts": {
-    "start": "react-scripts --openssl-legacy-provider start",
-    "build": "react-scripts --openssl-legacy-provider build",
+    "start": "react-scripts start",
+    "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
