@@ -1,48 +1,9 @@
-import React, { Component } from "react";
-import Fade from "react-reveal";
-
-class Footer extends Component {
-  render() {
-    if (!this.props.data) return null;
-
-    const networks = this.props.data.social.map(function (network) {
-      return (
-        <li key={network.name}>
-          <a href={network.url}>
-            <i className={network.className}></i>
-          </a>
-        </li>
-      );
-    });
-
+function Footer () {
     return (
-      <footer>
-        <div className="row">
-          <Fade bottom>
-            <div className="twelve columns">
-              <ul className="social-links">{networks}</ul>
-
-              <ul className="copyright">
-                <li>&copy; Copyright 2022 Matt Gainer</li>
-                <li>
-                  Design by{" "}
-                  <a title="Styleshout" href="http://www.styleshout.com/">
-                    Styleshout
-                  </a>
-                </li>
-              </ul>
-            </div>
-          </Fade>
-
-          <div id="go-top">
-            <a className="smoothscroll" title="Back to Top" href="#home">
-              <i className="icon-up-open"></i>
-            </a>
-          </div>
-        </div>
-      </footer>
-    );
-  }
+        <footer>
+            <p>&#169; Copyright 2022 Matt Gainer Design by Matt Gainer</p>
+        </footer>
+    )
 }
 
 export default Footer;
